# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Social App API
  description: API documentation for the Social App backend
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/users/me/profile:
    get:
      tags:
      - User Profile
      summary: Get authenticated user's profile
      description: Retrieves the profile of the authenticated user. Helper-specific
        fields are null for non-helpers.
      operationId: getMyProfile
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      responses:
        "200":
          description: Profile retrieved successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
          content:
            text/plain:
              example: Invalid or expired Firebase token
      security:
      - BearerAuth: []
    patch:
      tags:
      - User Profile
      summary: Update authenticated user's profile
      description: Updates the profile of the authenticated user with the provided
        fields
      operationId: updateProfile
      parameters:
      - name: Authorization
        in: header
        description: "Firebase authentication token in format: 'Bearer <token>'"
        required: true
        schema:
          type: string
        example: Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6...
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          description: Profile updated successfully
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Invalid or expired Firebase token
          content:
            text/plain:
              example: Invalid or expired Firebase token
        "400":
          description: Invalid request data
          content:
            application/json:
              example:
                error: Invalid profile data provided
      security:
      - BearerAuth: []
...
  /api/auth/profile:
    get:
      tags:
      - Authentication
      summary: Get current user profile
      description: Returns the profile of the currently authenticated user
      operationId: getProfile
      responses:
        "200":
          description: Profile retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
components:
  schemas:
    Address:
      type: object
      properties:
        addressid:
          type: integer
          format: int32
        streetNumber:
          type: integer
          format: int32
        street:
          type: string
        zipcode:
          type: integer
          format: int32
        neighbourhoodid:
          $ref: "#/components/schemas/Location"
...
    Badges:
      type: object
      properties:
        badgeid:
          type: integer
          format: int32
        badgeName:
          type: string
        badge_description:
          type: string
        isSpecialist:
          type: boolean
        xpReward:
          type: integer
          format: int32
        ratingid:
          $ref: "#/components/schemas/Ratings"
        badgeDescription:
          type: string
...
    Location:
      type: object
      properties:
        locationid:
          type: integer
          format: int32
        locationCenterPoint:
          type: integer
          format: int32
        locationRadius:
          type: integer
          format: int32
        neighbourhoodid:
          type: integer
          format: int32
        neighbourhoodName:
          type: string
    Ratings:
      type: object
      properties:
        ratingid:
          type: integer
          format: int32
        ratingReview:
          type: string
        totalXpLevel:
          type: integer
          format: int32
        currentGroup:
          type: string
...
    User:
      type: object
      properties:
        userid:
          type: integer
          format: int32
        firebaseUid:
          type: string
        emailVerified:
          type: boolean
        phoneVerified:
          type: boolean
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        gender:
          type: string
        addressid:
          $ref: "#/components/schemas/Address"
        badgeid:
          $ref: "#/components/schemas/Badges"
        ratingid:
          $ref: "#/components/schemas/Ratings"
        userType:
          type: string
        isAdmin:
          type: boolean
...
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        skills:
          type: array
          items:
            type: string
        empty:
          type: boolean
...